Add some fixes to make SDK usable in the wallet
